5.16 update

+++ src/main/java/edu/scau/pyx/ssp/controller/AdministratorController.java
@@ -8,6 +8,9 @@ import edu.scau.pyx.ssp.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.Map;
 
@@ -20,38 +23,71 @@ public class AdministratorController {
     @Autowired
     private SensitiveKeywordService sensitiveKeywordService;
 
+    @RequestMapping(value = "/signin",method = RequestMethod.POST)
+    public boolean signIn(@RequestBody SystemUser user, HttpSession session){
+        return userService.adminSignIn(user,session);
+    }
+
     @RequestMapping(value = "/updateuserlockstatus",method = RequestMethod.POST)
-    public boolean updateUserLockStatus(@RequestBody SystemUser user){
+    public boolean updateUserLockStatus(@RequestBody SystemUser user,HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return false;
+        }
         return userService.updateUserLocked(user.getId(),user.isLocked());
     }
 
     @RequestMapping(value = "/getsensitivekeyword", method = RequestMethod.GET)
-    public List<SensitiveKeyword> getSensitiveKeyword(@RequestParam long begin, @RequestParam long length){
+    public List<SensitiveKeyword> getSensitiveKeyword(@RequestParam long begin, @RequestParam long length,HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return null;
+        }
         return sensitiveKeywordService.getSensitiveKeyword(begin, length);
     }
 
     @RequestMapping(value = "/getsensitivekeywordnum", method = RequestMethod.GET)
-    public long getSensitiveKeywordNum(){
+    public long getSensitiveKeywordNum(HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return -1;
+        }
         return sensitiveKeywordService.getSensitiveKeywordNum();
     }
 
     @RequestMapping(value = "/insertsensitivekeyword",method = RequestMethod.POST)
-    public boolean insertSensitiveKeyword(@RequestBody Map<String,String> keywordMap){
+    public boolean insertSensitiveKeyword(@RequestBody Map<String,String> keywordMap,HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return false;
+        }
         return sensitiveKeywordService.insertSensitiveKeyword(keywordMap.get("keyword"));
     }
 
     @RequestMapping(value = "/deletesensitivekeyword",method = RequestMethod.DELETE)
-    public boolean deleteSensitiveKeyword(@RequestParam(value = "id",required = true) long id){
+    public boolean deleteSensitiveKeyword(@RequestParam(value = "id",required = true) long id,HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return false;
+        }
         return sensitiveKeywordService.deleteSensitiveKeyword(id);
     }
 
     @RequestMapping(value = "/getuserlist", method = RequestMethod.GET)
-    public List<UserListInfo> getUserList(@RequestParam long begin, @RequestParam long length){
+    public List<UserListInfo> getUserList(@RequestParam long begin, @RequestParam long length,HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return null;
+        }
         return userService.getUserList(begin,length);
     }
 
     @RequestMapping(value = "/getusernum", method = RequestMethod.GET)
-    public long getUserNum(){
+    public long getUserNum(HttpSession session){
+        SystemUser admin = (SystemUser) session.getAttribute("user");
+        if(admin==null||!admin.getRole().equals("admin")){
+            return -1;
+        }
         return userService.getUserNum();
     }
 }
